+++ crates/full-node/sov-sequencer/src/preferred/preferred_blob_sender.rs
@@ -22,7 +22,6 @@ use crate::{common::TxStatusBlobSenderHooks, TxStatusManager};
 /// Wrapper around [`BlobSender`] with preferred blob -specific logic.
 pub struct PreferredBlobSender<Da: DaService> {
     inner: Option<BlobSender<Da, TxStatusBlobSenderHooks<Da::Spec>, LedgerDb>>,
-    nb_of_concurrent_blob_submissions: Arc<AtomicUsize>,
     nb_of_concurrent_batch_blob_submissions: Arc<AtomicUsize>,
 }
 
@@ -38,13 +37,11 @@ impl<Da: DaService> PreferredBlobSender<Da> {
         blobs_sender_channel: broadcast::Sender<BlobExecutionStatus<Da::Spec>>,
         seq_role: SequencerRole,
     ) -> anyhow::Result<(Self, Option<JoinHandle<()>>)> {
-        let nb_of_concurrent_blob_submissions = Arc::new(AtomicUsize::new(0));
         let nb_of_concurrent_batch_blob_submissions = Arc::new(AtomicUsize::new(0));
         match seq_role {
             SequencerRole::PgSyncReplica | SequencerRole::DaOnlyReplica => Ok((
                 Self {
                     inner: None,
-                    nb_of_concurrent_blob_submissions,
                     nb_of_concurrent_batch_blob_submissions,
                 },
                 None,
@@ -66,15 +63,14 @@ impl<Da: DaService> PreferredBlobSender<Da> {
                     blob_processing_timeout,
                     Some(blobs_sender_channel),
                     blobs_to_send,
-                    nb_of_concurrent_blob_submissions.clone(),
                     nb_of_concurrent_batch_blob_submissions.clone(),
+                    Arc::new(AtomicUsize::new(0)),
                 )
                 .await?;
 
                 Ok((
                     Self {
                         inner: Some(inner),
-                        nb_of_concurrent_blob_submissions,
                         nb_of_concurrent_batch_blob_submissions,
                     },
                     Some(blob_sender_handle),
@@ -135,10 +131,6 @@ impl<Da: DaService> PreferredBlobSender<Da> {
             }
         }
         Ok(())
-    }
-
-    pub(crate) fn nb_of_in_flight_blobs(&self) -> Arc<AtomicUsize> {
-        self.nb_of_concurrent_blob_submissions.clone()
     }
 
     pub(crate) fn nb_of_in_flight_batch_blobs(&self) -> Arc<AtomicUsize> {